backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
sqlalchemy==2.0.41
psycopg[binary]==3.2.9
pgvector==0.2.5
python-multipart==0.0.20
pypdf==5.4.0
python-docx==1.1.2
openai==1.108.1
PyJWT==2.10.1
